v23: syncbase: add hooks for injecting fake system clock data
Note, this CL ended up larger than expected:
- I had to add a mutex around VClock.SystemClock
- This forced me to restructure how clients interact with
VClock.SystemClock to prevent races, which in turn led to
making the field private and adding accessors
- In adding accessors, I had to decide what kind of access
API to expose
- After looking at the existing modules accessing
VClock.SystemClock, I realized we'd benefit from an API of
the form SysClockVals{,IfNotChanged}
- Switching to this API involved updating the existing users
of VClock.SystemClock; the result is that the accessors
all follow the same, simple pattern and are easier to
reason about
- As a side effect, I discovered that syncVClock's GetTime
request didn't need to be sent within the VClockData
transaction, so I was able to simplify that code as well

This CL also fixes a bug that was introduced in my previous
refactoring CL (17806). Before CL 17806, syncVClock and
getDeltasFromPeer were each passed a "peer connInfo", and
each of these methods then proceeded to run a small bit of
duplicated code to call copyMemberInfo and populate the
mount tables part of the connInfo struct. In CL 17086 I
inadvertently removed this bit of code from syncVClock. In
this CL, I move the copyMemberInfo functionality to
syncer.go's syncerWork, avoiding the code duplication and
ensuring that both syncVClock and getDeltasFromPeer are
passed a fully populated connInfo struct.

Beyond adding the time injection and access hooks, I also
implemented a few tests that use them as a way to
demonstrate and test their usage.

MultiPart: 1/2

Change-Id: Ic1f9c7e5afe01b07c2388a82fa0b9a17c3ec6161