query: alpha impl of secondary indexes (for use by discovery)
Syncbase does not [yet?] support secondary indexes on tables; but
discovery needs them for InterfaceName.  That is, discovery
needs to know if a query is limited to a specific interface name.
This maps nicely to secondary indexes; as such, an alpha version
of secondary index support is provided.

Why alpha?
This implementation is missing two important features:
1. A single secondary index consisting of multiple fields.
2. A secondary index of any type other than string.

Impementation:
Datasources communicate secondary indexes to the query package
via a new function on the datasource.Table interface:
  GetIndexFields() []Index

Index is defined as:
  type Index struct {
    FieldName string
    Kind      vdl.Kind
  }

Discovery implements GetIndexFields as follows:
  return []datasource.Index{datasource.Index{
      FieldName: "v.InterfaceName", Kind: vdl.String}}

Scan now takes the following arg:
  Scan(idxRanges ...datasource.IndexRanges)
  (Note: key ranges are now just another index range, but
   for field of "k".)

idxRanges[0] will always be for the "k" (i.e., the key).
idxRanges[1] will be for "v.InterfaceName", since that is what
was the one and only field returned from GetIndexFields.

IndexRanges is defined as the following:
  type IndexRanges struct {
    FieldName    string
    Kind         vdl.Kind
    NilAllowed   bool //true if Scan should return nil idx values
    StringRanges *StringFieldRanges
    // TODO(jkline): add fields for other types of indexes.
  }

StringFieldRanges is defined as:
  type StringFieldRange struct {
    Start string
    Limit string
  }
  type StringFieldRanges []StringFieldRange

Discovery needs to look for the special case where a single
interface name is being queried.  This will be the case if
*all* of the following are true:
  1. idxRanges[1].NilAllowed == false
  2. len(idxRanges[1].StringRanges) == 1
  3. len(idxRanges[1].StringRanges[0].Start) > 0
  4. idxRanges[1].StringRanges[0].Limit == append(idxRanges[1].StringRanges[0].Start, 0)
If all of the above are true, discovery need only look for
the single interface name of idxRanges[1].StringRanges[0].Start.

The above logic is implemented in x/ref/lib/discovery/query.go:
func limitToSingleInterface(idxRange datasource.IndexRanges) (bool, string)

It is not part of the query package and its use as well as tests are left to disovery.
Having said that, I put in some Printf statements and limitToSingleInterface.

MultiPart: 1/2
Change-Id: If7c1f97a52d8cccc0459554fdc9b75e7c481ace0