v23: syncbase: more vclock tests
Implements tests for updating the virtual clock based on
communication with an NTP server. Also implements a test for
sysclock check frequency (unrelated to NTP).

As part of this CL, I made some changes to the fake NTP server:
- moved it under x/ref/services/syncbase/testutil (which
seems like a better home for it)
- added a flag to allow specifying its "start time"
- changed it to have the OS assign an available port to
listen on, and to print out the resulting host for
consumption by tests (i.e. by ExpectVar)

MultiPart: 1/2

Change-Id: Ibeda31de6f158e0ec066af57a08d5b623ebd1475