syncbase: Simplify ping pong benchmark
Store int32 instead of time.Time and print out directional average times.
We no longer need DevModeGetTime.

It looks like the parameters peerSyncInterval and watchPollInterval still
dominate these times.

I have chosen 500 iterations of ping pong to establish our baseline.

With 50 and 100, we get about 200ms of latency per round trip.
With 50 and 10, we get about 150ms of latency per round trip.
With 10 and 10, we get about 32ms.
With 5 and 5, we get about 26ms.

(With 1 and 1, we can get as low as 9ms for small values of b.N, but as
b.N increases, the number rises rapidly.)

To make comparisons more fair, we may want to
* reuse the same prefix (when writing to the table)
* decrease the values in parameters.go

Change-Id: I42b16d759a48c9499d91f015cb51a52e175b0b38