v23/syncbase: Add extra flags to syncbase ping pong benchmark
You can now add numSync and numGroup flags to the test runner in order
to vary the number of peers in the syncgroup as well as the number of
overall syncgroups.

This is going to be used to examine how latency changes as these
numbers vary.

For example, you can run with 2 syncgroups and 3 peers:
jiri go test -test.bench=BenchmarkPingPongPair -test.benchtime=5s
  -run XXX -numGroup=2 -numSync=3 -v23test-syncbase-debug-args "-vpath=*=0"

Change-Id: I23922f1740d7101538597acabd35c0c35cd81715