v23: Ensure security/openssl.c isn't built in go1.6
This is basically just ensuring that the build constraints in
openssl.c and ecdsa_openssl.go are identical.

Note that it's weird to trigger off go1.6, since presumably we
should ignore these files in go1.7 and later as well.  But this
seems good enough for now; perhaps we'll stop supporting go1.5 or
earlier at some point, and just remove these files completely.

Change-Id: I0bb50c26fb264b44346490a8a6d15d6310772abd