ref: Fix vom and vdl bugs wrt recursive types.
This CL started out as a fix for a vom.TypeDecoder bug.  The old
code would fail for recursive types if the wire type messages
weren't in a specific order.  E.g. given this type:

  type Recursive struct {
    X *Recursive
  }

The failure would occur if the message describing Recursive
occurred before the message describing *Recursive,
since *Recursive would enter the "unnamed type" codepath, which
would fail complaining that there is a cycle.  The fix is to
restructure the logic to ensure that pending is always populated
with a real pending type representing a given type id.

The fix exposed another bug in vdl uniqueTypeStr.  Cyclic types
where no type is named in the cycle would cause uniqueTypeStr to
infinite loop.  The fix is twofold: 1) disallow unnamed cyclic
types, and 2) break out of the loop if the type hasn't been
consed yet.  After the type has been consed and Type.unique is
set, we know that there aren't any unnamed cyclic loops, so
uniqueTypeStr will always terminate, and needs to actually return
fully unique strings.

Change-Id: I17ed2b16b63140c300e9feda2ef8d85c74205c63