v.io/v23/services/syncbase/nosql: Add device type to SyncgroupMemberInfo
This change replaces the "IsServer" bit in SyncgroupMemberInfo
with a device-type field, which can indicate "server", "normal",
or "leaf".

Simulations of blob migrations showed that without some way of
indicating that certain devices should almost never be the
destinations for migrations (such as a camera), unwelcome
migrations would occurs when collections of devices were offline
from their syncgroups' servers.

MultiPart: 2/2
Change-Id: Icfa66a9e34a2259d332bb83968c83b4f3cbd68ec