v23: New VDL / VOM Decoder.
This CL introduces vdl.Decoder, which is an interface that
provides DFS access to an encoded vdl value.  The experimental
vom.XDecoder implements this interface, which both simplifies the
code and provides a speedup.  Here are some benchmark numbers
using the new decoder, tested using a new VDL codegen (not
included in this CL).

                                old ns/op  new ns/op  delta
BenchmarkVomDecodeRequest       10499      2166       -79.37%
BenchmarkVomDecodeResponse      2072       589        -71.57%
BenchmarkVomDecoder1Customer    195984     193085     -1.48%
BenchmarkVomDecoder1000Customer 4737189    2676894    -43.49%

PROTO (for reference)
BenchmarkProtoDecodeRequest   1000000  1876 ns/op
BenchmarkProtoDecodeResponse  1000000  1033 ns/op

The new vom.XDecoder still uses the old vom.Decoder to decode
type messages, because it's simpler for now.  It's also
completely untested at the moment.

The purpose of checking in this experimental code is to unblock
other changes (e.g. the vdl.Value and reflect implementations of
VDLRead and Decoder), which are necessary for full codegen
support.  It's also easier to review as a smaller CL.

Change-Id: I72ead367e2ca275a3e4472b9f08877e7e5975989