Add VDLIsZero, and use it in VDLWrite.
This CL adds the VDLIsZero method, which is used to check whether
the method receiver is the VDL zero value.  This is used in
codegen to make zero struct field checking both simpler and faster.

Also simplified both VDLRead and VDLWrite codegen, to make things
more consistent and easier to understand.

MultiPart: 1/5
Change-Id: I381a8fe7672787fb2bacb835e81fda99b01ed21b