Add vdltest package, defining vdl test types and values.
The concept is that the vdltest package provides a variety of
test types, as well as lists of test entries.  Each entry
contains a pair of values, representing a conversion test from
the Source to a value of the Target type.  Our vdl/vom tests will
start using these test cases as the basis for our conversion and
protocol conformance tests.

This is a first cut at the package.  Many TODOs remain, wrt
covering more useful cases, and better pruning out of cases that
aren't as useful.

MultiPart: 1/3
Change-Id: Icdace687d9e976e1034b0009d67aa0e35288d8df