Update vdltest to include cyclic and native types.
The strategy for both of these is to include them in
vtype_manual.vdl and xtype_manual.vdl, which hold manually
defined types.  These are manually-defined because they are hard
to generate automatically, and easier to debug when defined
manually.  The test entries are still automatically-generated
based on the manual type definitions.

Also fixed some bugs wrt native types, exposed by the new tests.

MultiPart: 1/3
Change-Id: I81c7e7e5c969debf60ff4f4515d45b4d43ec727d