"Final fixes" for vdl/vom.  All tests pass with the new coders.
This is a seemingly random collection of small (but critical)
fixes.  Most of the issues are wrt reflection and RawBytes, which
isn't surprising.

The hardest bug to track down was the NumStarted%2 bug in
xEncoder, which resulted in us tracking the anylens incorrectly,
and also not encoding new types when we should.  After debugging
forever and poring through vom dumps, on a whim I tried flipping
the key/elem count, and it all worked.  In hindsight all the
symptoms really pointed at exactly this problem.

Change-Id: I062637537d73b0264b926effa56103e148ff07af