Change verror WireToNative behavior.
The old vdl/vom logic had a special-case when decoding into
interface{}; if we found a type that wasn't registered, we'd fall
back to creating a *vdl.Value.

The new logic returns an error in this case instead.  That
changes how verror.WireToNative behaves; using the old logic, any
types that weren't registered would result in *vdl.Value, which
the java side handles correctly.  The new logic would embed an
error into the verror params, which causes Java to behave
erratically.

Instead of embedding the error, we fall back to the same
special-case, where we just set the param to *vdl.Value instead.
This seems better since it provides more data to the user.

Change-Id: I706f75ce4d817ada7bed71c62fbf76bc3e3eccbb