TBR Change vom benchmarks so we don't overuse memory.
The old logic, when running a VomDecodeRepeated test, would
create a buffer with N encoded values, and decode each of those
values.  That creates huge contiguous buffers, which causes our
machines to start swapping and die.

The new logic creates a buffer with 2 encoded values.  Each time
we want to decode, we seek to the offset of the second value, and
decode from there.

Also changed the naming of the tests so it's easier to see what
we're testing.

Change-Id: Ia11c251a9e378e09c629bf615aa8b792dfa16465