security: Do not skip ACL checks when using same principal.
Syncbase is frequently run using the same principal (key pair) as
its clients. The default authorizer would skip all ACL checks in
this case, which is undesirable since it can e.g. allow the client
to put rows into a read-only collection, which would be rejected
by a remote peer when synced.

For this reason, this CL removes the short-circuit check for the
same public key and always checks the ACL.

MultiPart: 1/3
Change-Id: Idfcc470bd93f4d67fd8e4391a4f087704165e85f