v.io/v23/syncbase: call context.CancelFunc functions
In a previous cleanup, we made sure to call all context.CancelFunc
functions where we would otherwise leak memory.

The check in change 22953 is now more strict.  It enforces the spec of
Go's context package, which requires that the context.CancelFunc
be called, even if the context will be cancelled by other means.

This change fixes up the places in syncbase that the new check
reveals.

MultiPart: 1/2
Change-Id: I4b3671302f5916be1899342cdcacce70f3122b14