v.io/v23/security: fix context.CancelFunc leaks
The context.CancelFunc returned by calls like context.WithCancel()
etc, must be called to avoid a memory leak.

The leaks aren't not really important in these tests.  I'm fixing them mostly
to silence the context leak checker, so any output from the context leak
checker is worth looking at.  Also, it's probably a good thing so that if
anyone copies the code style, they'll see that the CancelFunc is called.

Change-Id: Ifaa3884f5aa3b59db5cc4ec59047d155fc2bff68